Move notifications feed helper
