+++ internal/notifications/notifier.go
@@ -6,6 +6,7 @@ import (
 	"log"
 	"sync"
 	ttemplate "text/template"
+	"time"
 
 	"github.com/arran4/goa4web/config"
 	"github.com/arran4/goa4web/core/templates"
@@ -127,4 +128,32 @@ func (n *Notifier) notifyAdmins(ctx context.Context, et *EmailTemplates, nt *str
 		}
 	}
 	return nil
+}
+
+// NotificationPurgeWorker periodically removes old read notifications.
+func (n *Notifier) NotificationPurgeWorker(ctx context.Context, interval time.Duration) {
+	if n.Queries == nil {
+		return
+	}
+	ticker := time.NewTicker(interval)
+	defer ticker.Stop()
+	for {
+		select {
+		case <-ticker.C:
+			if err := n.Queries.PurgeReadNotifications(ctx); err != nil {
+				log.Printf("purge notifications: %v", err)
+			}
+		case <-ctx.Done():
+			return
+		}
+	}
+}
+
+// sendInternalNotification stores an internal notification for the user.
+func (n *Notifier) sendInternalNotification(ctx context.Context, userID int32, path, msg string) error {
+	return n.Queries.InsertNotification(ctx, dbpkg.InsertNotificationParams{
+		UsersIdusers: userID,
+		Link:         sql.NullString{String: path, Valid: path != ""},
+		Message:      sql.NullString{String: msg, Valid: msg != ""},
+	})
 }
